Solution to Calculate Average Temperature of a Week

+++ Calculate Average Temperature of a Week@@67877d61dc107486ad5c274d/code.c
@@ -4,7 +4,7 @@ struct Day
 {
     char d[20];
     char n[20];
-    float t;
+    int t;
 };
 int main()
 {
@@ -15,7 +15,7 @@ int main()
     {
         scanf("%s",e[a].d);
         scanf("%s",e[a].n);
-        scanf("%f",&e[a].t);
+        scanf("%d",&e[a].t);
     }
     for(a=0;a<7;a++)
     {
